uimenu.c
The menu is Emterpreted-Asynced. This code keeps the menu UI separate from the emscripten main loop, so that the main loop does not itself need to be Emterpreted.

There's also a hack to cover Emscripten SDL1's lack of SDL_WaitEvent, which may have made sense before Emterpret-Async existed, but doesn't now.